ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
redis==5.0.1
boto3==1.34.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
transformers==4.36.0
torch==2.2.0
spacy==3.7.2
pandas==2.1.3
camelot-py[cv]==0.11.0
langdetect==1.0.9
python-dotenv==1.0.0
httpx==0.25.2